% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_01.R
\docType{package}
\name{Covid19CasesPlot-package}
\alias{Covid19CasesPlot-package}
\title{Plot COVID19 Patterns Using Data from John Hopkins University.}
\description{
Retrieve latest COVID-19 data provided by John Hopkins University (Baltimore, US), 
obtained from GitHub: "CSSEGISandData/COVID-19/". For more info, please
see: \url{https://systems.jhu.edu/research/public-health/ncov/}. 
Visualize the trend of new cases and deaths due to COVID-19 infection
in a country or state of interest. Points indicate 
the number of new cases or deaths measured day by day. Trendlines are 
computed by 'loess'.
}
\examples{
library(Covid19CasesPlot)
library(ggplot2)
x1 <- fetch_latest_covid19_data()
get_available_dates(x1)
get_available_geo_areas(x1)[1:10]
p1 <- plot_covid19_data(covid19.data = x1, geo.area = "Italy")
p1

}
\references{
Info and resources are available at the following URLs:
\enumerate{
  \item{\strong{JHU website}: \url{https://systems.jhu.edu/research/public-health/ncov/} }
  \item{\strong{GitHub library}: \url{https://github.com/dami82/Covid19CasesPlot} }
  \item{\strong{Daily Charts}: \url{https://systems.jhu.edu/research/public-health/ncov/} }
 }
}
