% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_01.R
\name{fetch_latest_covid19_data}
\alias{fetch_latest_covid19_data}
\title{Fetch Latest COVID19 data.}
\usage{
fetch_latest_covid19_data(US.only = FALSE)
}
\arguments{
\item{US.only}{logical, shall US-only be downloaded. If FALSE, global
data are downloaded.}
}
\value{
list including three elements:
\describe{
  \item{cse.df}{data.frame including COVID19 cumulative cases data}
  \item{dth.df}{data.frame including COVID19 cumulative death data}
  \item{US.only}{logical, value of the US.only argument}
}
}
\description{
Retrieve latest data provided by 
John Hopkins University (Baltimore, US), 
from GitHub: "CSSEGISandData/COVID-19/".
}
\examples{
library(Covid19CasesPlot)
x <- fetch_latest_covid19_data()

}
