% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_01.R
\name{get_available_dates}
\alias{get_available_dates}
\title{Get All Available Dates.}
\usage{
get_available_dates(covid19.data)
}
\arguments{
\item{covid19.data, }{list returned by fetch_latest_covid19_data(). 
It includes worldwide or US COVID19 infection data.}
}
\value{
list, including 4 elements:
  * first.day: date of the first data point
  * last.day: date of the latest data point
  * window: integer, number of days in the window
  * data.window: integer, number of data points
}
\description{
Retrieve some summary data about the dates/time points included in the 
dataset.
}
\examples{
library(Covid19CasesPlot)
x <- fetch_latest_covid19_data()
get_available_dates(x)

}
