% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_01.R
\name{plot_covid19_data}
\alias{plot_covid19_data}
\title{Plot Numbers of New COVID19 Cases and Deaths.}
\usage{
plot_covid19_data(covid19.data, geo.area, y.limit = NULL,
  dth.scale.k = NULL)
}
\arguments{
\item{covid19.data, }{list returned by fetch_latest_covid19_data(). 
It includes worldwide COVID19 infection data.}

\item{geo.area, }{string indicate the country or the US-state of interest. 
For example, 'US', 'Italy' (country), or "Alabama" (US-state).}

\item{y.limit, }{integer of length 1. This is the maximum value used
in the y axis (New cases). If NULL, it is automatically set to the 
95th percentile.}

\item{dth.scale.k, }{integer of length 1. This is the coefficient 
used to scale the secondary y-axis (deaths counts). If NULL, it
is automatically computed. Typically, reasonable numbers range between 3 (US, 
United Kingdom, Italy), and 30 (Germany, Japan).}
}
\value{
a ggplot-class object.
}
\description{
Visualize the trend of new cases and deaths due to COVID-19 infection
in a country of interest. A ggplot2-object is returned. Points indicate 
the number of new cases or deaths measured day by day. Trendlines are 
computed by 'loess'.
}
\examples{
library(Covid19CasesPlot)
x <- fetch_latest_covid19_data()
plot_covid19_data(x, "US")

}
